# readium-api/utils/swagger/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Readium API",
    "description": "Handsome backend devs"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/dev\r/post": {
      "post": {
        "tags": [
          "Dev"
        ],
        "summary": "Create a post (DEV)",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverImage": {
                    "type": "file"
                  },
                  "title": {
                    "type": "string",
                    "example": "Post Title"
                  },
                  "content": {
                    "type": "string",
                    "example": "Text only content"
                  },
                  "textEditorContent": {
                    "type": "string",
                    "example": "Text Editor Content"
                  },
                  "description": {
                    "type": "string",
                    "example": "Description"
                  },
                  "isPublished": {
                    "type": "string",
                    "example": "false"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/following/posts": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get following writers' posts",
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/users/collections\r/posts": {
      "post": {
        "tags": [
          "Collection"
        ],
        "summary": "Add post to collection",
        "description": "",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostToCollection"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collection"
        ],
        "summary": "Delete post from collection",
        "description": "",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostToCollection"
              }
            }
          }
        }
      }
    },
...
    "/posts/me": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get my published posts",
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/posts/user/{userId}": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get user id's published posts",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{id}/cover-image": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get post's cover image",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "View a post",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Post"
        ],
        "summary": "Delete a post",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/posts/": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get posts",
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{id}/unpublish": {
      "put": {
        "tags": [
          "Post"
        ],
        "summary": "Endpoint to unpublish the post",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PostToCollection": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "example": "post id"
          },
          "collectionId": {
            "type": "string",
            "example": "collection id"
          }
        },
        "required": [
          "postId",
          "collectionId"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}